+++ UI/src/app/modules/dataset/components/dataset/dataset.component.ts
@@ -1,16 +1,16 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {DatasetService} from '../../../core/services/dataset/dataset.service';
-import {DatasetDto} from '../../../core/services/dataset/dataset.dto';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { DatasetService } from '../../../core/services/dataset/dataset.service';
+import { DatasetDto } from '../../../core/services/dataset/dataset.dto';
 
-import {Observable} from 'rxjs';
-import {DatasetVersionDto} from '../../../core/services/dataset/dataset-version.dto';
-import {DatasetVersionService} from '../../../core/services/dataset/dataset-version.service';
-import {DatasetComponentSelectionModel} from './dataset-component-selection-model';
-import {DatasetVariableDto} from '../../../core/services/dataset/dataset-variable.dto';
-import {DatasetVariableService} from '../../../core/services/dataset/dataset-variable.service';
-import {map} from 'rxjs/operators';
-import {ValidationRunConfigService} from '../../../../pages/validate/service/validation-run-config.service';
-import {AuthService} from '../../../core/services/auth/auth.service';
+import { Observable } from 'rxjs';
+import { DatasetVersionDto } from '../../../core/services/dataset/dataset-version.dto';
+import { DatasetVersionService } from '../../../core/services/dataset/dataset-version.service';
+import { DatasetComponentSelectionModel } from './dataset-component-selection-model';
+import { DatasetVariableDto } from '../../../core/services/dataset/dataset-variable.dto';
+import { DatasetVariableService } from '../../../core/services/dataset/dataset-variable.service';
+import { map, tap } from 'rxjs/operators';
+import { ValidationRunConfigService } from '../../../../pages/validate/service/validation-run-config.service';
+import { AuthService } from '../../../core/services/auth/auth.service';
 
 
 @Component({
@@ -27,12 +27,12 @@ export class DatasetComponent implements OnInit {
   selectableDatasetVariables$: Observable<DatasetVariableDto[]>;
 
   selectableDatasetVersionsObserver = {
-    next: versions => this.onSelectableVersionsNext(versions),
+    next: (versions: DatasetVersionDto[]) => this.onSelectableVersionsNext(versions),
     complete: () => this.onSelectableVersionsComplete()
   }
 
   selectableDatasetVariablesObserver = {
-    next: variables => this.onSelectableVariablesNext(variables),
+    next: (variables: DatasetVariableDto[]) => this.onSelectableVariablesNext(variables),
     complete: () => this.onSelectableVariablesComplete()
   }
 
@@ -43,6 +43,8 @@ export class DatasetComponent implements OnInit {
   datasetSelectorId: string;
   versionSelectorId: string;
   variableSelectorId: string;
+  newerVersionExists = false;
+  newestVersionId: number;
 
   constructor(private datasetService: DatasetService,
               private datasetVersionService: DatasetVersionService,
@@ -53,45 +55,46 @@ export class DatasetComponent implements OnInit {
 
 
   ngOnInit(): void {
-    this.allDatasets$ = this.datasetService.getAllDatasets(true)
+    this.allDatasets$ = this.datasetService.getAllDatasets(true);
 
     this.validationConfigService.listOfSelectedConfigs.subscribe(configs => {
       if (configs.filter(config => config.datasetModel.selectedDataset?.short_name === 'ISMN').length !== 0
         && this.selectionModel.selectedDataset?.short_name !== 'ISMN') {
         this.datasets$ = this.allDatasets$
           .pipe(map<DatasetDto[], DatasetDto[]>(datasets => {
-            return this.sortById(datasets.filter(dataset => dataset.pretty_name !== 'ISMN'));
+            return datasets.filter(dataset => dataset.pretty_name !== 'ISMN');
           }));
       } else if (configs.filter(config => config.datasetModel.selectedDataset?.user).length == configs.length - 1 && !this.selectionModel.selectedDataset?.user) {
         this.datasets$ = this.allDatasets$
           .pipe(map<DatasetDto[], DatasetDto[]>(datasets => {
-            return this.sortById(datasets.filter(dataset => !dataset.user));
+            return datasets.filter(dataset => !dataset.user);
           }));
       } else {
         this.datasets$ = this.allDatasets$
           .pipe(map<DatasetDto[], DatasetDto[]>(datasets => {
-            return this.sortById(datasets);
+            return datasets;
           }));
       }
     });
 
 
-    this.selectableDatasetVersions$ = this.sortObservableById(
-      this.datasetVersionService.getVersionsByDataset(this.selectionModel.selectedDataset.id), false);
+    this.selectableDatasetVersions$ = this.datasetVersionService.getVersionsByDataset(this.selectionModel.selectedDataset.id).pipe(
+      tap(datasetVersions => this.checkIfNewerVersionExists(datasetVersions))
+    );
 
-    this.selectableDatasetVariables$ = this.sortObservableById(
-      this.datasetVariableService.getVariablesByDataset(this.selectionModel.selectedDataset.id));
+    this.selectableDatasetVariables$ = this.datasetVariableService.getVariablesByVersion(this.selectionModel.selectedVersion.id);
 
     this.setSelectorsId();
   }
 
-  private updateSelectableVersionsAndVariableAndEmmit(): void {
+   updateSelectableVersionsAndVariableAndEmmit(): void {
     if (this.selectionModel.selectedDataset === undefined || this.selectionModel.selectedDataset.versions.length === 0) {
       return;
     }
 
-    this.selectableDatasetVersions$ = this.sortObservableById(
-      this.datasetVersionService.getVersionsByDataset(this.selectionModel.selectedDataset.id), false);
+    this.selectableDatasetVersions$ = this.datasetVersionService.getVersionsByDataset(this.selectionModel.selectedDataset.id).pipe(
+      tap(versions => this.checkIfNewerVersionExists(versions))
+    );
 
 
     this.selectableDatasetVersions$.subscribe(this.selectableDatasetVersionsObserver);
@@ -103,10 +106,9 @@ export class DatasetComponent implements OnInit {
   }
 
   private onSelectableVersionsComplete(): void {
-    this.selectableDatasetVariables$ = this.sortObservableById(
-      this.datasetVariableService.getVariablesByDataset(this.selectionModel.selectedDataset.id));
+    this.selectableDatasetVariables$ = this.datasetVariableService.getVariablesByVersion(this.selectionModel.selectedVersion.id);
 
-    this.selectableDatasetVariables$.subscribe(this.selectableDatasetVariablesObserver)
+    this.selectableDatasetVariables$.subscribe(this.selectableDatasetVariablesObserver);
     this.setSelectorsId();
   }
 
@@ -119,27 +121,14 @@ export class DatasetComponent implements OnInit {
     this.setSelectorsId();
   }
 
-  onDatasetChange(): void {
-    this.updateSelectableVersionsAndVariableAndEmmit();
-  }
-
-  onVersionChange(): void {
+  onVersionChange(versions: DatasetVersionDto[]): void {
+    this.checkIfNewerVersionExists(versions);
     this.changeDataset.emit(this.selectionModel);
   }
 
-  sortById(listOfElements): any {
-    return listOfElements.sort((a, b) => {
-      return a.id < b.id ? 1 : -1;
-    });
-  }
-
-  sortObservableById(observableOfListOfElements: Observable<any>, minToMax = true): Observable<any> {
-    return observableOfListOfElements.pipe(map((data) => {
-      data.sort((a, b) => {
-        return minToMax ? (a.id > b.id ? 1 : -1) : (a.id < b.id ? 1 : -1);
-      });
-      return data;
-    }));
+  checkIfNewerVersionExists(versions: DatasetVersionDto[]): void {
+    this.newerVersionExists = Math.max(...versions.map(version => version.id)) > this.selectionModel.selectedVersion.id
+    this.newestVersionId = Math.max(...versions.map(version => version.id));
   }
 
   setSelectorsId(): void {