+++ validator/admin/dataset_variable_custom.py
@@ -2,9 +2,9 @@ from django.contrib.admin.options import ModelAdmin
 
 
 class DatasetVariableAdmin(ModelAdmin):
-    list_display = ('id', 'short_name', 'dataset')
+    list_display = ('id', 'short_name', 'version')
 
-    def dataset(self, obj):
+    def version(self, obj):
         if len(obj.variables.all()):
             return obj.variables.all()[0]
         else: