+++ validator/validation/readers.py
@@ -217,7 +217,6 @@ def create_reader(dataset, version) -> GriddedNcTs:
     if dataset.short_name == globals.SMAP_L2:
         reader = GriddedNcOrthoMultiTs(folder_name,
                                        ioclass_kws={'read_bulk': True})
-        # and version.short_name == globals.SMAP_V9_AM_PM
 
     if (dataset.short_name == globals.SMAP_L3 and version.short_name ==
             'SMAP_V9_AM_PM'):
@@ -285,16 +284,6 @@ def adapt_timestamp(reader, dataset, version):
     elif dataset.short_name == globals.SMAP_L2:
         tadapt_kwargs = {
             'base_time_field': 'acquisition_time',
-            'base_time_reference': '2000-01-01T12:00:00',
-            'base_time_units': 's',
-            'time_offset_fields': None,
-            'time_units': None,
-        }
-
-    elif (dataset.short_name == globals.SMAP_L3 and version.short_name ==
-          globals.SMAP_V6_PM):
-        tadapt_kwargs = {
-            'base_time_field': 'tb_time_seconds',
             'base_time_reference': '2000-01-01T12:00:00',
             'base_time_units': 's',
             'time_offset_fields': None,