+++ release-notes.md
@@ -1,3 +1,19 @@
+QA4SM v3.1.0.1 - Release notes 2025-05-13
+=======================================================
+# Fixes
+1. Missing filters added
+
+QA4SM v3.1.0 - Release notes 2025-05-12
+=======================================================
+# Updates
+1. Automatic account verification via email
+2. Handling of validation config reloading with outdated datasets
+3. Outdated dataset versions removed
+4. SMAP3 data v9 added 
+5. Variable now stored in version model, not dataset model
+6. 'Deleted validation' model added, stores metadata on removed validations
+7. User data can be used as spatial reference
+
 QA4SM v3.0.4 - Release notes 2025-03-24
 =======================================================
 # Updates